server/pyproject.toml

[project]
name = "soluna-chat-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "sqlmodel==0.0.22",
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.3.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "websockets==12.0",
    "python-dotenv==1.0.1",
    "pytest==8.3.2",
    "anyio==4.4.0",
    "httpx==0.27.2",
]
